still trying to fix ban

+++ backend/controllers/user.js
@@ -125,10 +125,12 @@ export const banUser = async (req, res) => {
     try {
         const { userId } = req.params
 
-        await db.query(`DELETE FROM seen_messages WHERE messageId IN (SELECT messageId FROM messages WHERE senderId = ?)`, [userId]),
-        await db.query(`DELETE FROM messages WHERE senderId = ?`, [userId]),
-        await db.query(`DELETE FROM room_participants WHERE userId = ?`, [userId]),
-        await db.query(`DELETE FROM users WHERE userId = ?`, [userId])
+        await Promise.all([
+            await db.query(`DELETE FROM seen_messages WHERE messageId IN (SELECT messageId FROM messages WHERE senderId = ?) AND seenBy = ?`, [userId, userId]),
+            await db.query(`DELETE FROM messages WHERE senderId = ?`, [userId]),
+            await db.query(`DELETE FROM room_participants WHERE userId = ?`, [userId]),
+            await db.query(`DELETE FROM users WHERE userId = ?`, [userId])
+        ])
 
         res.status(200).json('Banned')
     } catch (error) {
